added utils_tests.rs and moved AST printing to it
